# openapi.yaml
openapi: 3.0.0
info:
  title: Atende API
  version: 1.0.0
  description: API for the Atende application
paths:
...
  /api/cruds/{id}:
    get:
      summary: Get a CRUD by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single CRUD
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crud'
        '404':
          description: CRUD not found
    put:
      summary: Update a CRUD by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrudInput'
      responses:
        '200':
          description: CRUD updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Crud'
        '400':
          description: Bad request - Invalid configuration or duplicate name
        '404':
          description: CRUD not found
    delete:
      summary: Delete a CRUD by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: CRUD deleted
        '404':
          description: CRUD not found
...
components:
...
  schemas:
...
    Crud:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          description: Unique identifier for the CRUD
        title:
          type: string
          description: Display title
        icon:
          type: string
          description: Material icon name
        resource:
          type: string
          description: Resource name for permissions
        endpoint:
          type: string
          description: API endpoint path
        config:
          type: object
          description: JSON configuration object
          additionalProperties: true
        active:
          type: boolean
          description: Whether the CRUD is active
      required:
        - name
        - title
        - resource
        - endpoint
        - config
    CrudInput:
      type: object
      properties:
        name:
          type: string
          description: Unique identifier for the CRUD
        title:
          type: string
          description: Display title
        icon:
          type: string
          description: Material icon name
        resource:
          type: string
          description: Resource name for permissions
        endpoint:
          type: string
          description: API endpoint path
        config:
          type: object
          description: JSON configuration object
          additionalProperties: true
        active:
          type: boolean
          description: Whether the CRUD is active
      required:
        - name
        - title
        - resource
        - endpoint
        - config
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT